Better hotplug support for the plugin
Add a layer of indirection between the sensor classes and UsbDevice so the device can be rescanned if it gets unplugged.

Hot unplug should now result in only a single log line being emitted, and then another one if the device is replugged, after which, the device should continue to work as normal.

Devices that were not plugged in at plugin initialization (ie: when Fan Control is started) will still not show up, due to limitations of the plugin interface.