Extend to work with multiple fan connections
Currently only supports 2 fans, but should be able to add a 3rd easily enough.

PWM duty cycle can be set independently for each fan, but frequency is shared among all of them, due to limitations of the hardware.

Change register interface revision to 1.0. Now that there are multiple fans, I decided to keep the registers as-is, even though the PWM frequency being grouped with the first fan's duty cycle is a little weird. SO the register defs are no longer subject to compatibility-breaking changes without bumping the major number.