Support for third fan per USB device
Had to use PCINT2 for the input instead of INT2, which would have been a lot more consistent with the other two fan connections, because the INT2 interrupt was flakey on all 4 board varieties I tried. It looked like there was some crosstalk with the PWM signal, which was weird, given that polling the input line instead of using the interrupt did not have the same problem.

Switch the input pin configurations to having the internal pull-up resistor enabled, since the common case will be that at least one of them will be left floating, thus prone to generate spurious interrupts.

Add a check on the RPM calculation if it overflows a 16-bit int. No, that fan probably isn't running at 65K RPM, but the afore mentioned crosstalk can make it look that way.